+++ packages/cadl-ranch-specs/http/azure/core/lro/rpc/mockapi.ts
@@ -1,4 +1,4 @@
-import { passOnSuccess, mockapi, json } from "@azure-tools/cadl-ranch-api";
+import { passOnSuccess, json, MockRequest } from "@azure-tools/cadl-ranch-api";
 import { ScenarioMockApi } from "@azure-tools/cadl-ranch-api";
 
 export const Scenarios: Record<string, ScenarioMockApi> = {};
@@ -6,25 +6,74 @@ export const Scenarios: Record<string, ScenarioMockApi> = {};
 let generationPollCount = 0;
 
 Scenarios.Azure_Core_Lro_Rpc_longRunningRpc = passOnSuccess([
-  mockapi.post("/azure/core/lro/rpc/generations:submit", (req) => {
-    req.expect.containsQueryParam("api-version", "2022-12-01-preview");
-    req.expect.bodyEquals({ prompt: "text" });
-    generationPollCount = 0;
-    return {
+  {
+    uri: "/azure/core/lro/rpc/generations:submit",
+    method: "post",
+    request: {
+      body: { prompt: "text" },
+      params: {
+        "api-version": "2022-12-01-preview",
+      },
+    },
+    response: {
       status: 202,
       headers: {
-        "operation-location": `${req.baseUrl}/azure/core/lro/rpc/generations/operations/operation1`,
+        "operation-location": `/azure/core/lro/rpc/generations/operations/operation1`,
       },
       body: json({ id: "operation1", status: "InProgress" }),
-    };
-  }),
-  mockapi.get("/azure/core/lro/rpc/generations/operations/operation1", (req) => {
-    req.expect.containsQueryParam("api-version", "2022-12-01-preview");
-    const response =
-      generationPollCount > 0
-        ? { id: "operation1", status: "Succeeded", result: { data: "text data" } }
-        : { id: "operation1", status: "InProgress" };
-    generationPollCount += 1;
-    return { status: 200, body: json(response) };
-  }),
+    },
+    handler: (req: MockRequest) => {
+      req.expect.containsQueryParam("api-version", "2022-12-01-preview");
+      req.expect.bodyEquals({ prompt: "text" });
+      generationPollCount = 0;
+      return {
+        status: 202,
+        headers: {
+          "operation-location": `${req.baseUrl}/azure/core/lro/rpc/generations/operations/operation1`,
+        },
+        body: json({ id: "operation1", status: "InProgress" }),
+      };
+    },
+    kind: "MockApiDefinition",
+  },
+  {
+    uri: "/azure/core/lro/rpc/generations/operations/operation1",
+    method: "get",
+    request: {
+      params: {
+        "api-version": "2022-12-01-preview",
+      },
+    },
+    response: {
+      status: 200,
+      body: json({ id: "operation1", status: "InProgress" }),
+    },
+    handler: lroHandler,
+    kind: "MockApiDefinition",
+  },
+  {
+    uri: "/azure/core/lro/rpc/generations/operations/operation1",
+    method: "get",
+    request: {
+      params: {
+        "api-version": "2022-12-01-preview",
+      },
+    },
+    response: {
+      status: 200,
+      body: json({ id: "operation1", status: "Succeeded", result: { data: "text data" } }),
+    },
+    handler: lroHandler,
+    kind: "MockApiDefinition",
+  },
 ]);
+
+function lroHandler(req: MockRequest) {
+  req.expect.containsQueryParam("api-version", "2022-12-01-preview");
+  const response =
+    generationPollCount > 0
+      ? { id: "operation1", status: "Succeeded", result: { data: "text data" } }
+      : { id: "operation1", status: "InProgress" };
+  generationPollCount += 1;
+  return { status: 200, body: json(response) };
+}