+++ packages/cadl-ranch/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @azure-tools/cadl-ranch
 
+## 0.14.7
+
+### Patch Changes
+
+- f1614e1: bump dependency to typespec 0.61.0
+- 352934c: Backport the changes from typespec repository to migrate the scenarios to new model
+- Updated dependencies [f1614e1]
+- Updated dependencies [352934c]
+  - @azure-tools/cadl-ranch-expect@0.15.5
+  - @azure-tools/cadl-ranch-api@0.4.7
+
 ## 0.14.6
 
 ### Patch Changes